requirements.txt
flask==0.12.2
requests==2.18.4
pycryptodome==3.6.1
